Fixed warning
Conversion warning (from 'float' to 'double').

+++ tests/src/wx_metric/wx_metric_distance.cpp
@@ -24,7 +24,7 @@ void wx_metric_distance_max(const wxgraph<float> *G) {
 }
 
 void wx_metric_distance_mean(const wxgraph<float> *G) {
-	double d = networks::metrics::distance::max_distance(G);
+	float d = networks::metrics::distance::max_distance(G);
 	cout << floatpointout(d) << endl;
 }
